# services/ekstraksi/openapi.yaml
openapi: 3.1.0
info:
  title: OCR NPWP API
  description: 'OCR service for Indonesian NPWP (tax ID card) documents; ServiceOCR in the pipeline. **Async
    pipeline:** the orchestrator POSTs /v1/ekstraksi/jobs after guardrails passed and gets 202; this service
    runs OCR in the background, stores the result, POSTs a stage callback to the orchestrator, and hands
    the job to the structuring service. **Legacy contract:** generate-request-id -> extract-ocr -> get-ocr-result
    runs the whole chain synchronously and stays until the orchestrator has moved to the async flow. The
    raw OCR step is also exposed as /v1/ekstraksi/extract. All endpoints except /health require an X-API-Key
    header.


    ## Conventions (the same for every nilam-ocr service)


    **Authentication.** Every endpoint except `/health` and `/ready` requires the header `X-API-Key`.

    A missing or wrong key answers `401`. The services are reachable only from inside the cluster.


    **Envelope.** Every JSON response, success or error, has the same shape:

    `{status_code, status_desc, message, data, errors, request_id}`. `status_code` always equals the HTTP

    status. On success `errors` is null; on error `data` is null.


    **Errors.** `errors` is machine-readable: `VALIDATION_ERROR` for `422`, otherwise the same text as

    `message`. `400` = the request or the document is unusable (do not retry unchanged). `401` = API key.

...

...

...


...
    no

...

    without it the service generates one.


...

...

...

...

...

...

...
servers:
...
paths:
...
  /v1/ekstraksi/jobs:
    post:
      tags:
      - Pipeline
      summary: Start the pipeline for a document (OCR stage)
      description: '**Step 2 of the pipeline, asynchronous: the entry point of the OCR -> structuring
        -> scoring chain.** Call it after `POST /v1/guardrails/check` answered `passed: true`.


        Records the job (`ocr.jobs`, idempotent per request_id), answers **202 immediately**, then in
        the background: reads the document (`file`, or downloads `file_url`), runs OCR, stores the result
        (`ocr.results`), POSTs the `OCR` callback, and hands the job to the structuring service, which
        hands it to scoring. The caller does nothing more: it receives one callback per stage, and the
        final result in the `SCORING` callback (see *Webhooks*).


        **Document.** Send `file` (multipart) or `file_url`, exactly one. JPEG, PNG or PDF, at most 5
        MB. `file_url` is downloaded in the background, so make a presigned URL live longer than the worst
        queueing time; an expired or unreachable URL becomes a `FAILED` job, not a `4xx`.


        **Idempotency.** The same request_id again answers `202` with `duplicate: true` and does not run
        OCR twice, unless the earlier attempt `FAILED`, in which case it is run again.'
      operationId: submitOcrJob
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Body_submitOcrJob'
        required: true
      responses:
        '202':
          description: The job was accepted (or already existed)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobAcceptedResponse'
              examples:
                accepted:
                  summary: New job
                  value:
                    status_code: 202
                    status_desc: Accepted
                    message: Accepted
                    data:
                      request_id: REQ_9cb01af2-493d-446d-b191-af120333f6d0
                      stage: OCR
                      status: PROCESSING
                      duplicate: false
                    request_id: REQ_9cb01af2-493d-446d-b191-af120333f6d0
                duplicate:
                  summary: 'Same request_id sent again: nothing is re-run, `status` is the existing job''s
                    status'
                  value:
                    status_code: 202
                    status_desc: Accepted
                    message: Accepted
                    data:
                      request_id: REQ_9cb01af2-493d-446d-b191-af120333f6d0
                      stage: OCR
                      status: DONE
                      duplicate: true
                    request_id: REQ_9cb01af2-493d-446d-b191-af120333f6d0
        '400':
          description: Neither or both of file / file_url, or `guardrails` is not a JSON object
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                status_code: 400
                status_desc: Bad Request
                message: Send exactly one of file or file_url
                errors: Send exactly one of file or file_url
        '401':
          description: Missing or invalid X-API-Key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                status_code: 401
                status_desc: Unauthorized
                message: Invalid or missing API key
                errors: Invalid or missing API key
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                status_code: 422
                status_desc: Unprocessable Entity
                message: 'body.request_id: Field required'
                errors: VALIDATION_ERROR
      security:
      - APIKeyHeader: []
  /v1/ekstraksi/jobs/{request_id}:
    get:
      tags:
      - Pipeline
      summary: Status and result of the OCR stage
      description: Status of this stage only, and its raw OCR result once `DONE`. The orchestrator normally
        learns the status from the callback; use this to reconcile after a missed callback, or to debug.
        The later stages have the same endpoint on their own service (`/v1/structuring/jobs/{request_id}`,
        `/v1/scoring/jobs/{request_id}`).
      operationId: getOcrJob
      security:
      - APIKeyHeader: []
      parameters:
      - name: request_id
        in: path
        required: true
        schema:
          type: string
          title: Request Id
      responses:
        '200':
          description: The job exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OcrJobStatusResponse'
              examples:
                processing:
                  summary: Still running
                  value:
                    status_code: 200
                    status_desc: OK
                    message: Success
                    data:
                      request_id: REQ_9cb01af2-493d-446d-b191-af120333f6d0
                      stage: OCR
                      created_at: '2026-09-18T04:00:00+00:00'
                      status: PROCESSING
                      updated_at: '2026-09-18T04:00:00+00:00'
                    request_id: REQ_9cb01af2-493d-446d-b191-af120333f6d0
                done:
                  summary: Finished
                  value:
                    status_code: 200
                    status_desc: OK
                    message: Success
                    data:
                      request_id: REQ_9cb01af2-493d-446d-b191-af120333f6d0
                      stage: OCR
                      created_at: '2026-09-18T04:00:00+00:00'
                      status: DONE
                      result:
                        engine: paddle
                        model: PP-OCRv6_medium_det+PP-OCRv6_medium_rec
                        elapsed_ms: 412.5
                        full_text: 'DIREKTORAT JENDERAL PAJAK

                          NPWP : 12.345.678.9-012.345

                          BUDI SANTOSO'
                        blocks:
                        - text: DIREKTORAT JENDERAL PAJAK
                          confidence: 0.9985
                          bbox:
                            x1: 175
                            y1: 247
                            x2: 887
                            y2: 308
                          page: 0
                        - text: 'NPWP : 12.345.678.9-012.345'
                          confidence: 0.9992
                          bbox:
                            x1: 35
                            y1: 389
                            x2: 637
                            y2: 437
                          page: 0
                        - text: BUDI SANTOSO
                          confidence: 0.9773
                          bbox:
                            x1: 33
                            y1: 480
                            x2: 483
                            y2: 523
                          page: 0
                      updated_at: '2026-09-18T04:00:01+00:00'
                    request_id: REQ_9cb01af2-493d-446d-b191-af120333f6d0
                failed:
                  summary: 'Failed: same `error_message` as the FAILED callback. Resubmitting the request_id
                    runs it again'
                  value:
                    status_code: 200
                    status_desc: OK
                    message: Success
                    data:
                      request_id: REQ_9cb01af2-493d-446d-b191-af120333f6d0
                      stage: OCR
                      created_at: '2026-09-18T04:00:00+00:00'
                      status: FAILED
                      error_message: ekstraksi OCR model is unavailable
                      updated_at: '2026-09-18T04:00:03+00:00'
                    request_id: REQ_9cb01af2-493d-446d-b191-af120333f6d0
        '401':
          description: Missing or invalid X-API-Key
          content:
            application/json:
              example:
                status_code: 401
                status_desc: Unauthorized
                message: Invalid or missing API key
                errors: Invalid or missing API key
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: No OCR job for this request_id
          content:
            application/json:
              example:
                status_code: 404
                status_desc: Not Found
                message: 'No OCR job found for request_id: REQ_9cb01af2-493d-446d-b191-af120333f6d0'
                errors: 'No OCR job found for request_id: REQ_9cb01af2-493d-446d-b191-af120333f6d0'
                request_id: REQ_9cb01af2-493d-446d-b191-af120333f6d0
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              example:
                status_code: 422
                status_desc: Unprocessable Entity
                message: 'path.request_id: Field required'
                errors: VALIDATION_ERROR
                request_id: REQ_9cb01af2-493d-446d-b191-af120333f6d0
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
...
    Body_submitOcrJob:
      properties:
        request_id:
          type: string
          title: Request Id
          description: request_id minted by the orchestrator
          examples:
          - REQ_9cb01af2-493d-446d-b191-af120333f6d0
        document_type:
          type: string
          title: Document Type
          description: Document type chosen by the client. Only `npwp` is supported
          default: npwp
          examples:
          - npwp
        guardrails:
          anyOf:
          - type: string
          - type: 'null'
          title: Guardrails
          description: 'The `data` object of `POST /v1/guardrails/check`, serialised as a JSON string.
            Forwarded down the chain: scoring uses `document.confidence`, and the final result returns
            it unchanged. Optional, but without it the scoring model works with one input missing'
          examples:
          - '{"passed": true, "reason": null, "document": {"verdict": "accepted", "confidence": 0.9821,
            "n_pages": 1, "n_approve": 1, "n_reject": 0}, "pages": [{"page_index": 0, "proba_approve":
            0.9821, "proba_reject": 0.0179, "verdict": "accepted"}]}'
        file:
          anyOf:
          - type: string
            format: binary
          - type: string
          - type: 'null'
          title: File
          description: Document image (JPEG/PNG/PDF). Omit when sending file_url.
        file_url:
          anyOf:
          - type: string
          - type: 'null'
          title: File Url
          description: URL this service fetches the document image from (e.g. a presigned MinIO GET).
            Omit when uploading file.
      type: object
      required:
      - request_id
      title: Body_submitOcrJob
    BoundingBox:
      properties:
        x1:
          type: integer
          title: X1
          description: Left
          examples:
          - 190
        y1:
          type: integer
          title: Y1
          description: Top
          examples:
          - 320
        x2:
          type: integer
          title: X2
          description: Right
          examples:
          - 760
        y2:
          type: integer
          title: Y2
          description: Bottom
          examples:
          - 366
      type: object
      required:
      - x1
      - y1
      - x2
      - y2
      title: BoundingBox
...
    ErrorResponse:
      properties:
        status_code:
          type: integer
          title: Status Code
          description: Same as the HTTP status code
          examples:
          - 400
        status_desc:
          type: string
          title: Status Desc
          description: Reason phrase of `status_code`
          examples:
          - Bad Request
        message:
          type: string
          title: Message
          description: Human-readable cause; safe to log
          examples:
          - Uploaded file is empty
        data:
          type: 'null'
          title: Data
          description: Always null on error
        errors:
          type: string
          title: Errors
          description: Machine-readable cause. `VALIDATION_ERROR` for 422 (branch on this, the message
            names a different field every time); for every other status it repeats `message`
          examples:
          - Uploaded file is empty
        request_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Request Id
          description: Present when the error is tied to a known request_id
          examples:
          - REQ_9cb01af2-493d-446d-b191-af120333f6d0
      type: object
      required:
      - status_code
      - status_desc
      - message
      - errors
      title: ErrorResponse
...
    JobAccepted:
      properties:
        request_id:
          type: string
          title: Request Id
          description: Echo of the submitted request_id
          examples:
          - REQ_9cb01af2-493d-446d-b191-af120333f6d0
        stage:
          type: string
          enum:
          - OCR
          - STRUCTURING
          - SCORING
          title: Stage
          description: Pipeline stage that accepted the job
          examples:
          - OCR
        status:
          type: string
          enum:
          - PROCESSING
          - DONE
          - FAILED
          title: Status
          description: '`PROCESSING` for a newly accepted job. When `duplicate` is true: the current status
            of the job that already exists for this request_id'
          examples:
          - PROCESSING
        duplicate:
          type: boolean
          title: Duplicate
          description: 'true: this request_id was already accepted before and the work is NOT run again
            (idempotent). Exception: a job whose status is `FAILED` is re-claimed and run again, and answers
            `duplicate: false`'
          default: false
          examples:
          - false
      type: object
      required:
      - request_id
      - stage
      - status
      title: JobAccepted
    JobAcceptedResponse:
      properties:
        status_code:
          type: integer
          title: Status Code
          description: Same as the HTTP status code
          default: 202
          examples:
          - 202
        status_desc:
          type: string
          title: Status Desc
          description: Reason phrase of `status_code`
          default: Accepted
          examples:
          - Accepted
        message:
          type: string
          title: Message
          description: Human-readable outcome
          default: Accepted
          examples:
          - Accepted
        errors:
          type: 'null'
          title: Errors
          description: Always null on success
        request_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Request Id
          description: 'The request_id this response belongs to: the one sent in the form / body / path,
            otherwise the `X-Request-ID` header, otherwise one generated by the service'
          examples:
          - REQ_9cb01af2-493d-446d-b191-af120333f6d0
        data:
          $ref: '#/components/schemas/JobAccepted'
      type: object
      required:
      - data
      title: JobAcceptedResponse
...
    OcrJobStatus:
      properties:
        request_id:
          type: string
          title: Request Id
          description: request_id of the pipeline run
          examples:
          - REQ_9cb01af2-493d-446d-b191-af120333f6d0
        stage:
          type: string
          const: OCR
          title: Stage
          description: Always `OCR` on this service
          default: OCR
          examples:
          - OCR
        status:
          type: string
          enum:
          - PROCESSING
          - DONE
          - FAILED
          title: Status
          description: '`PROCESSING` -> `DONE` | `FAILED`. A `FAILED` job may go back to `PROCESSING`
            when resubmitted'
          examples:
          - DONE
        error_message:
          anyOf:
          - type: string
          - type: 'null'
          title: Error Message
          description: Why the stage failed; null unless `status` is `FAILED`. Same text as the FAILED
            callback
          examples:
          - null
        created_at:
          type: string
          title: Created At
          description: When the job was first accepted (ISO 8601, UTC)
          examples:
          - '2026-09-18T04:00:00+00:00'
        updated_at:
          type: string
          title: Updated At
          description: Last status change (ISO 8601, UTC)
          examples:
          - '2026-09-18T04:00:01+00:00'
        result:
          anyOf:
          - $ref: '#/components/schemas/OcrResult'
          - type: 'null'
          description: Raw OCR result once `status` is `DONE`; null otherwise
      type: object
      required:
      - request_id
      - status
      - created_at
      - updated_at
      title: OcrJobStatus
    OcrJobStatusResponse:
      properties:
        status_code:
          type: integer
          title: Status Code
          description: Same as the HTTP status code
          default: 200
          examples:
          - 200
        status_desc:
          type: string
          title: Status Desc
          description: Reason phrase of `status_code`
          default: OK
          examples:
          - OK
        message:
          type: string
          title: Message
          description: Human-readable outcome
          default: Success
          examples:
          - Success
        errors:
          type: 'null'
          title: Errors
          description: Always null on success
        request_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Request Id
          description: 'The request_id this response belongs to: the one sent in the form / body / path,
            otherwise the `X-Request-ID` header, otherwise one generated by the service'
          examples:
          - REQ_9cb01af2-493d-446d-b191-af120333f6d0
        data:
          $ref: '#/components/schemas/OcrJobStatus'
      type: object
      required:
      - data
      title: OcrJobStatusResponse
    OcrResult:
      properties:
        engine:
          type: string
          title: Engine
          description: OCR backend that read the document (`EKSTRAKSI_BACKEND`)
          examples:
          - paddle
        model:
          anyOf:
          - type: string
          - type: 'null'
          title: Model
          description: Model identity reported by the OCR model service; null when it does not report
            one
          examples:
          - PP-OCRv6_medium_det+PP-OCRv6_medium_rec
        elapsed_ms:
          type: number
          title: Elapsed Ms
          description: Time spent in the OCR model, milliseconds
          examples:
          - 412.5
        full_text:
          type: string
          title: Full Text
          description: All blocks joined by newline, in reading order
          examples:
          - 'DIREKTORAT JENDERAL PAJAK

            NPWP : 12.345.678.9-012.345

            BUDI SANTOSO'
        blocks:
          items:
            $ref: '#/components/schemas/TextBlock'
          type: array
          title: Blocks
          description: Text lines in reading order (top to bottom)
      type: object
      required:
      - engine
      - elapsed_ms
      - full_text
      - blocks
      title: OcrResult
...
    TextBlock:
      properties:
        text:
          type: string
          title: Text
          description: One OCR text line, as read
          examples:
          - 'NPWP : 12.345.678.9-012.345'
        confidence:
          type: number
          maximum: 1.0
          minimum: 0.0
          title: Confidence
          description: Recognition score of this line
          examples:
          - 0.9992
        bbox:
          anyOf:
          - $ref: '#/components/schemas/BoundingBox'
          - type: 'null'
          description: Upright box around the line, in pixels of the ORIENTATION-CORRECTED image the OCR
            model worked on. The model straightens and unwarps the card first, so these are NOT coordinates
            in the uploaded photo
        page:
          type: integer
          minimum: 0.0
          title: Page
          description: 0-based page number (multi-page PDF)
          default: 0
          examples:
          - 0
      type: object
      required:
      - text
      - confidence
      title: TextBlock
  securitySchemes:
    APIKeyHeader:
      type: apiKey
      description: API key for this service
      in: header
      name: X-API-Key
...